discovery keyring init and refactoring
