+++ examples/test/keyrings/test_i_hierarchical_keyring_example.py
@@ -3,7 +3,7 @@
 """Test suite for the hierarchical keyring example."""
 import pytest
 
-from ...src.keyrings.hierarchical_keyring import encrypt_and_decrypt_with_keyring
+from ...src.keyrings.hierarchical_keyring_example import encrypt_and_decrypt_with_keyring
 
 pytestmark = [pytest.mark.examples]
